fix: split up key into i18nPrefix and routePrefix
This is needed to properly distinguish between nested routes and translation files.
While we cannot expect all resource routes to be top-level, we do expect all translation
files translating models to be.
Thus, to properly resolve both locations from one key, we need to split it up into
a translation key and a routing key, where the former can be derived from the latter by
simply taking the last segment.