refactor: pagination
Use proper bootstrap styling to have the styling more in line with
current bootstrap themes.
Refactor page size control to be individual links; to get rid of the form.

Refactor CanAutoSort trait to be injected via Macros into the query builder
instead.

The Macro needs to be set on both the base query builder as well as the
eloquent builder, because the eloquent builder will not return the value
returned by a dynamic function call, instead returning itself.
This means `autoPaginate` can only work when living directly on the eloquent
builder.