feat: sleek setup command
Added a new console command `sleek:setup` to automate the setup process for the Sleek package.
The command checks if `bootstrap` and `bootstrap-icons` npm packages are installed and offers to install them if they are not.
Implemented a function to verify the presence of the correct `@import` statement in the `app.scss` file.
If the SCSS import is incorrect or missing, the command prompts the user to add the correct import statement and appends it to the `app.scss` file if confirmed.